fix(training-agent): wire list_accounts handler + reuse shared cursor codec
Addresses three blockers from the independent expert review on #3110:

1. **Idempotency_key.** The `sync_three_accounts` setup step on the new
   storyboard omitted `idempotency_key`, failing the build-time
   storyboard-idempotency lint. Adds the conventional
   `$generate:uuid_v4#<storyboard>_<phase>_<step>` alias.

2. **Cursor codec duplication.** Triage shipped private
   `encodeAccountCursor` / `decodeAccountCursor` in account-handlers.ts
   with format `offset:N`. PR #3095 (now merged) generalized the codec
   to a kind-prefixed pair. Moves the shared codec to
   `server/src/training-agent/pagination.ts` (avoids the
   task-handlers ↔ account-handlers circular import the in-place
   re-export would create) and replaces the local helpers with
   `encodeOffsetCursor('accounts', n)` / `decodeOffsetCursor('accounts', c)`.
   list_creatives and get_signals handlers move to the shared module
   with no behavior change.

3. **Bootstrap-pattern deviation note.** The seed-DAG documented at
   storyboard-schema.yaml does not include accounts; this storyboard
   uses `sync_accounts` as setup rather than `controller_seeding: true`.
   Adds explicit narrative explaining why and pointing at a future
   `seed_account` controller scenario.

Also fixes:

- **Wire-shape leak.** `accountStateToWire()` was passing `account.brand`
  through verbatim, including a `name` operational hint that brand-ref.json
  forbids (additionalProperties: false). Strips to declared fields only
  (`domain`, optional `brand_id`); display name and advertiser fields are
  still derived from the operational hint but never round-trip on the
  wire. Compliance fixture pool's brand entries cleaned up to match.

- **Framework registration.** Triage added `handleListAccounts` to
  `HANDLER_MAP` but never registered the tool through the
  framework-server's `accounts:` block, so the SDK never advertised
  `list_accounts` and the storyboard's pagination steps were skipped
  (no-such-tool). Wires it alongside `sync_accounts`.

- **Doc-parity.** Adds the new storyboard to both
  `docs/building/conformance.mdx` and `docs/building/compliance-catalog.mdx`
  per the lint that gates the build.

Verified: negative-test flips `has_more` to false on the agent — first-page
assertion fires with `Expected true, got false`. Reverted, clean run
restored. 4/4 storyboard steps pass.
