+++ server/src/training-agent/task-handlers.ts
@@ -19,6 +19,7 @@ import { mergeSeedProduct } from '@adcp/client/testing';
 import { isDatabaseInitialized, getPool } from '../db/client.js';
 import { createLogger } from '../logger.js';
 import type { TrainingContext, CatalogProduct, MediaBuyState, PackageState, SignalActivationState, CreativeState, CreativeManifest, ToolArgs, ListReference, PackageTargeting } from './types.js';
+import { encodeOffsetCursor, decodeOffsetCursor } from './pagination.js';
 import type {
   Product,
   Proposal,
@@ -2104,35 +2105,6 @@ export async function handleListCreatives(args: ToolArgs, ctx: TrainingContext) 
       return base;
     }),
   };
-}
-
-// Opaque offset-encoded cursors for in-memory paginated reads. Real backends
-// would carry stable resource keys; an offset is sufficient for the training
-// agent because the underlying iteration order is stable for every read this
-// powers (Map insertion order, static catalog order). base64url keeps the
-// token URL-safe and visibly opaque. The `kind` prefix scopes a cursor to a
-// specific list endpoint so a caller can't move a list_creatives cursor onto
-// list_creative_formats and accidentally land at a meaningful offset.
-function encodeOffsetCursor(kind: string, offset: number): string {
-  return Buffer.from(`${kind}:offset:${offset}`).toString('base64url');
-}
-
-// Returns null when the cursor is present but malformed (or scoped to a
-// different list endpoint). The caller MUST surface INVALID_REQUEST in that
-// case — silently restarting from offset 0 would teach a sloppy pattern
-// (corrupt cursors duplicate items the caller already saw). An absent
-// cursor returns 0 (start of pagination).
-function decodeOffsetCursor(kind: string, cursor: string | undefined): number | null {
-  if (!cursor) return 0;
-  try {
-    const decoded = Buffer.from(cursor, 'base64url').toString('utf8');
-    const m = new RegExp(`^${kind}:offset:(\\d+)$`).exec(decoded);
-    if (!m) return null;
-    const n = Number.parseInt(m[1], 10);
-    return Number.isFinite(n) && n >= 0 ? n : null;
-  } catch {
-    return null;
-  }
 }
 
 function encodeCreativeCursor(offset: number): string {
